feat: Implement dual-exchange live WebSocket ingestion (Binance & Alpaca) with fractional volume models

+++ src/main/java/com/quant/engine/model/TimeBar.java
@@ -20,6 +20,6 @@ public record TimeBar(
         double low,
         double close,
         double vwap,
-        long totalVolume
+        double totalVolume
 ) {
 }
